Move contents of buffer.h into format.h/format.inl
